+++ R/generate_rules.R
@@ -60,8 +60,10 @@ generate_rules <- function(X, ite_std, ntrees_rf, ntrees_gbm, node_size, max_nod
   }
 
   treelist_RF <- inTrees::RF2List(forest)
-  rules_RF <- extract_rules(treelist_RF, X, ntrees_rf, max_depth,
-                            ite_std, max_decay, type_decay)
+  rules_RF <- extract_rules(treelist_RF, X, ntrees_rf, max_depth)
+  M_rf_initial <- length(rules_RF)
+  rules_RF <- prune_rules(rules_RF, X, ite_std, max_decay, type_decay)
+  M_rf_pruned <- length(rules_RF)
 
   # Gradient Boosting
   dist <- ifelse(is.numeric(ite_std), "gaussian", "bernoulli")
@@ -83,8 +85,10 @@ generate_rules <- function(X, ite_std, ntrees_rf, ntrees_gbm, node_size, max_nod
 
 
   treelist_GB <- inTrees::GBM2List(model1_GB, X)
-  rules_GB <- extract_rules(treelist_GB, X, ntrees_gbm, max_depth,
-                            ite_std, max_decay, type_decay)
+  rules_GB <- extract_rules(treelist_GB, X, ntrees_gbm, max_depth)
+  M_rf_initial <- length(rules_GB)
+  rules_GB <- prune_rules(rules_GB, X, ite_std, max_decay, type_decay)
+  M_rf_pruned <- length(rules_GB)
 
 
   rules_list <- c(rules_RF, rules_GB)